change parseInt to parseFloat to allow operations with decimal

+++ js/scripts.js
@@ -21,8 +21,8 @@ function divide(number1, number2) {
 $(document).ready(function() {
     $("form#add").submit(function(event) {
       event.preventDefault();
-      const number1 = parseInt($("#add1").val());
-      const number2 = parseInt($("#add2").val());
+      const number1 = parseFloat($("#add1").val());
+      const number2 = parseFloat($("#add2").val());
       
       const result = add(number1, number2);
       $("#output").text(result);
@@ -30,8 +30,8 @@ $(document).ready(function() {
 
       $("form#subtract").submit(function(event) {
         event.preventDefault();
-        const number1 = parseInt($("#subtract1").val());
-        const number2 = parseInt($("#subtract2").val());
+        const number1 = parseFloat($("#subtract1").val());
+        const number2 = parseFloat($("#subtract2").val());
         
         const result = subtract(number1, number2);
         $("#output").text(result);
@@ -40,8 +40,8 @@ $(document).ready(function() {
 
         $("form#multiply").submit(function(event) {
           event.preventDefault();
-          const number1 = parseInt($("#multiply1").val());
-          const number2 = parseInt($("#multiply2").val());
+          const number1 = parseFloat($("#multiply1").val());
+          const number2 = parseFloat($("#multiply2").val());
           
           const result = multiply(number1, number2);
           $("#output").text(result);
@@ -49,12 +49,12 @@ $(document).ready(function() {
 
           $("form#divide").submit(function(event) {
             event.preventDefault();
-            const number1 = parseInt($("#divide1").val());
-            const number2 = parseInt($("#divide2").val());
+            const number1 = parseFloat($("#divide1").val());
+            const number2 = parseFloat($("#divide2").val());
             
             const result = divide(number1, number2);
             $("#output").text(result);
           });
   });
-  
+
